requirements.txt
Pillow==7.1.2
pymupdf==1.16.18
